src/third_party/blink/web_tests/platform/mac/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-multicol-expected.txt

This is a testharness.js-based test.
FAIL fieldset multicol assert_equals: expected "18px" but got "90px"
Harness: the test ran to completion.
